add comments to handlers, change like handler to prevent error if a user who has already liked it tries to like again

+++ handlers/delete.py
@@ -1,14 +1,18 @@
 from handler import Handler
 import models
 
+#allows creator of post to delete it
 class DeletePost(Handler):
     def post(self, id):
         blog = models.BlogPost.by_id(int(id))
         blog_user_id = str(blog.author_id)
         cookie_id = str(self.read_secure_cookie("id"))
-        if blog_user_id == cookie_id:
-            models.BlogPost.by_id(int(id)).delete()
-            self.redirect('/')
+        if cookie_id:
+            if blog_user_id == cookie_id:
+                models.BlogPost.by_id(int(id)).delete()
+                self.redirect('/')
+            else:
+                self.response.set_status(404, message="not permitted")
+                self.render('404_error.html')
         else:
-            self.response.set_status(404, message="not permitted")
-            self.render('404_error.html')
+            self.redirect('/login')
